Rewrite checksumming code, including the addition of a new I3Frame version (6), which is currently off by default pending an ack from blaufuss.
For v5 frames, the 16-bit CCITT CRC is now computed using a lookup table
instead of boost's "optimal" code, which is between a few percent and
twice as fast, depending on optimization.

For v6, frames, the 16-bit CCITT CRC is replaced by a CRC32-C. This is
a better checksum to begin with and exists in many more optimized
implementations that the CCITT checksum (for instance, it is implemented
in hardware on new Intel CPUs). Even on machines where it is not, a fallback
implementation taken from zlib implementation is included which is several
times faster than the CCITT code, making checksumming completely subdominant
(< 20% of total time without SSE4.2, not measurable with SSE4.2), instead of
being the dominant factor in serialization/deserialization time.

Merged from:	boost_test r1032-1038


git-svn-id: http://code.icecube.wisc.edu/icetray/projects/icetray/trunk@1156 579d543e-85e3-4578-a5e8-1e6a993b7602