Remove the (unused) naked-streambuf constructor, fixing the build with Boost 1.55.
Background: boost::interprocess::buffer stream and friends now inherit privately from their underlying streambufs, and since access is not considered when resolving overloads, this makes the portable_binary_iarchive constructor ambiguous. If for some reason you need to bind an archive to a free-standing streambuf, wrap it in a dummy iostream.

git-svn-id: http://code.icecube.wisc.edu/icetray/projects/icetray/trunk@1752 579d543e-85e3-4578-a5e8-1e6a993b7602