reset ROOT's SIGPIPE signal handler
root installs several of its own signal handlers (documented in
TUnixSystem::Init()). if `ctest` crashes or is killed while running a python
test, that test may continue. if it then tries to write to stdout/stderr,
(which was held by `ctest`) the file handle is gone, and this triggers a
SIGPIPE. because root's signal handler is sooooooooo smart, it also tries
to write to this missing filehandle, which in turn triggers a SIGPIPE,
causing an infinite loop.

hopefully this is all that's needed to stop that bs.



git-svn-id: http://code.icecube.wisc.edu/icetray/projects/icetray/trunk@2466 579d543e-85e3-4578-a5e8-1e6a993b7602